+++ tests/web/WebDriverTest.php
@@ -630,25 +630,4 @@ class WebDriverTest extends TestsForBrowsers
         $this->module->seeInField('foo', 'bar baz');
     }
 
-    public function testSelectOptionValueSelector()
-    {
-        $this->module->amOnPage('/form/select_selectors');
-
-        $this->module->selectOption('age', ['value' => '20']);
-        $this->module->seeOptionIsSelected('age', '21');
-
-        $this->module->selectOption('age', ['value' => '21']);
-        $this->module->seeOptionIsSelected('age', '20');
-    }
-
-    public function testSelectOptionTextSelector()
-    {
-        $this->module->amOnPage('/form/select_selectors');
-
-        $this->module->selectOption('age', ['text' => '20']);
-        $this->module->seeOptionIsSelected('age', '20');
-
-        $this->module->selectOption('age', ['text' => '21']);
-        $this->module->seeOptionIsSelected('age', '21');
-    }
 }